implementing reordering logic

+++ core/scope_savings_goals/domain/requests/reorder.go
@@ -2,6 +2,6 @@ package requests
 
 type Reorder struct {
 	ID   int64 `json:"id"`
-	From int64 `json:"from"`
-	To   int64 `json:"to"`
+	From int   `json:"from"`
+	To   int   `json:"to"`
 }
